cleaned + refactored (methods extracted to new file)
